+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "Twinkle" pattern configuration option for "Color Assignment".
 - "Twinkle" pattern configuration option for "Speed Deviation Chance".
 - "Twinkle" pattern configuration option for "Speed Deviation Multiplier".
+- New pattern "Pew".
+- "Pew" pattern configuration option for "Reverse Animation".
+- "Pew" pattern configuration option for "Burst Length".
 
 ## 0.5.0 - 2026-06-08